stability/style improvements: - add lat/lng nan check in MapLibreMapView.kt - removed containerWidth check in FoldableFlightScreen.kt, again - update readme with library information - apply param to api DataRequest.kt; use dynamic year for IataParserData.kt - use int for readSingle and usages -> .toInt()

+++ app/src/main/java/com/tuxy/airo/viewmodel/StandardDataViewModel.kt
@@ -30,7 +30,7 @@ class StandardDataViewModel(flightDataDao: FlightDataDao, id: String) : ViewMode
     init {
         viewModelScope.launch(Dispatchers.IO) {
             val job = viewModelScope.launch(Dispatchers.IO) { // TODO is this really the best way?
-                flightData.value = flightDataDao.readSingle(id) ?: FlightData()
+                flightData.value = flightDataDao.readSingle(id.toInt()) ?: FlightData()
             }
             job.join()
         }
